gringotts/requirements.txt

-i https://pypi.org/simple
alembic==1.7.7
anyio==3.5.0; python_full_version >= '3.6.2'
asgiref==3.5.0; python_version >= '3.7'
asyncpg==0.25.0
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3.5'
commonmark==0.9.1
coverage[toml]==6.3.2; python_version >= '3.7'
dnspython==2.2.1; python_version >= '3.6' and python_version < '4.0'
eventlet==0.33.0
fastapi==0.75.2
greenlet==2.0.0a2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.12.0; python_version >= '3.6'
httpcore==0.13.7; python_version >= '3.6'
httptools==0.4.0
httpx==1.0.0b0
idna==3.3; python_version >= '3.5'
iniconfig==1.1.1
mako==1.2.0; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
packaging==21.3; python_version >= '3.6'
pluggy==1.0.0; python_version >= '3.6'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.9.0
pyparsing==3.0.8; python_full_version >= '3.6.8'
pytest-asyncio==0.18.3
pytest-cov==3.0.0
pytest==7.1.1; python_version >= '3.7'
rfc3986[idna2008]==1.5.0
rich==12.2.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy[asyncio]==1.4.35
starlette==0.17.1; python_version >= '3.6'
tomli==2.0.1; python_full_version >= '3.6.0'
typing-extensions==4.2.0; python_version >= '3.7'
uvicorn==0.17.6
uvloop==0.16.0
pika
python-jose
python-multipart
psycopg2-binary

opentelemetry-sdk==1.11.1
opentelemetry-instrumentation-fastapi==0.30b1
opentelemetry-instrumentation-requests==0.30b1
opentelemetry-instrumentation-logging==0.30b1
opentelemetry-instrumentation-asyncpg==0.30b1
opentelemetry-instrumentation-pika==0.30b1
opentelemetry-instrumentation-digma==0.5.16
python-dotenv
